+++ daemons/attrd/attrd_utils.c
@@ -17,7 +17,7 @@
 #include <sys/types.h>
 
 #include <crm/crm.h>
-#include <crm/common/ipcs_internal.h>
+#include <crm/common/ipc_internal.h>
 #include <crm/common/mainloop.h>
 
 #include "pacemaker-attrd.h"
