+++ daemons/controld/controld_messages.c
@@ -18,7 +18,7 @@
 #include <crm/common/xml.h>
 #include <crm/cluster/internal.h>
 #include <crm/cib.h>
-#include <crm/common/ipcs_internal.h>
+#include <crm/common/ipc_internal.h>
 
 #include <pacemaker-controld.h>
